airflow/requirements.txt
apache-airflow-providers-postgres>=6.2.0
apache-airflow-providers-celery>=3.0.0
apache-airflow-providers-redis>=3.0.0
dbt-core==1.9.0
dbt-postgres==1.9.0
psycopg2-binary==2.9.9
requests==2.31.0